chore: add notification hook dir and move notification hooks
