+++ src/Navigation/MainNavigator.tsx
@@ -1,6 +1,6 @@
 import useAuth from 'hooks/auth/useAuth'
 import useProfile from 'hooks/auth/useProfile'
-import useNotification from 'hooks/independent/useNotification'
+import useNotification from 'hooks/notification/useNotification'
 import { useAppNavigation } from 'hooks/useAppNavigation'
 import { Routes } from 'libs/navigation'
 import { getPin } from 'libs/pin'
